pyproject.toml
[project]
name = "fastapiproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "asyncpg>=0.31.0",
    "fastapi>=0.134.0",
    "fastapi-users[sqlalchemy]>=15.0.4",
    "imagekitio>=5.2.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "uvicorn[standard]>=0.41.0",
]
